:finnadie: working on magnetic fields

+++ test/test_unit_cell.py
@@ -5,8 +5,8 @@ from mmag.unit_cell.cell import Cuboid
 
 class TestCubicCell(unittest.TestCase):
     def setUp(self):
-        position = np.array([0.0, 0.0, 0.0])
-        delta = np.array([1.0, 1.0, 1.0])
+        position = np.array([0.0, 0.0, 0.0], dtype=np.float64)
+        delta = np.array([1.0, 1.0, 1.0], dtype=np.float64)
         self.init_obj = Cuboid(position, delta)
 
     def test_input_1(self):
